Made tests pass for exception messages & not for stack trace
Some tests are matching with regular expressions which are not
present in the exception message, but present in the stack
trace of the exception object, e.g. qr/illegal/ matches the
stack trace of the exception, but not with the message, message
contains "Illegal".
Now, since stringification of exception objects contain only message,
so these tests should have regular expression that match exception
message.